Fixed UIComponent.Transparency (int -> float)
Woops, should be a float!

+++ Entities/Components/Internal/IInterfaceComponent.cs
@@ -12,7 +12,7 @@ namespace Andromeda.Entities.Components.Internal
         int ZIndex { get; set; }
         bool Visible { get; set; }
         UITransform Transform { get; }
-        int Transparency { get; set; }
+        float Transparency { get; set; }
         Color Color { get; set; }
 
         Vector2f AbsoluteSize { get; }
